refactor: move footer related components to the same level
