+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
 
 import Header from '@/components/Header/Header';
-import Footer from '@/components/Footer/Footer';
+import Footer from '@/components/footer/Footer';
 import { AuthProvider } from '@/context/AuthContext';
 const geistSans = Geist({
 	variable: '--font-geist-sans',
